Refactor: Extract classes Collect and Execute
+ Added Collect.cs
+ Added Execute.cs
