+++ Botkit/Microsoft.BotKit/Conversation/Execute.cs
@@ -1,0 +1,21 @@
+﻿// Copyright(c) Microsoft Corporation.All rights reserved.
+// Licensed under the MIT License.
+
+namespace Microsoft.BotKit.Conversation
+{
+    /// <summary>
+    /// Execute Class.
+    /// </summary>
+    public class Execute
+    {
+        /// <summary>
+        /// Gets or Sets the Script.
+        /// </summary>
+        public string Script { get; set; }
+
+        /// <summary>
+        /// Gets or Sets the Thread.
+        /// </summary>
+        public string Thread { get; set; }
+    }
+}
